Update AUCClassificationPerformanceEvaluator_2class.java
Update the calculation of PAUC, sPAUC, HoldoutAUC based on the new version of PAUC paper [1], which considers the event of score tie. The updated code fragments were copied from the BasicAUCImblanacedPerformanceEvaluator.java. I chose to update the code here because I found that the BasicAUCImblanacedPerformanceEvaluator.java is not calculating the PAUC with sliding window but calculating the AUC of the model from the beginning of the data stream to the current time step. i.e. the computation at the last time step of the finite stream is the AUC of the model to the whole data stream.

[1] Brzezinski, D., & Stefanowski, J. (2017). Prequential AUC: properties of the area under the ROC curve for data streams with concept drift. Knowledge and Information Systems, 52, 531-562.